Change I/O spec to simplify kernel-side PMA, add crude I/O prototype
It is much easier to use PMA for everything in the kernel but the old
model doesn't work well with it. The current model uses some fields to
mark an "empty" entry, eliminating the need to synchronize tail/head
between kernel and userspace.

The request entries have been sized up to 64 bytes so it doesn't cross
page boundaries.