Add dux library for managing memory reservations.
The library allows a form of "overcommit" that doesn't rely on
page-faults and OOM killers. A library or program can reserve a range of
virtual addresses which won't be allocated but which will be tracked by
the dux library. Pages can then safely be mapped into this space without
conflicting with other mappings. This can also be used to find unreserved
ranges.