Use OpenSBI, boot in S-Mode + map kernel into virtual memory
It's super hacky right now but it works.

The boot process goes as follows now:

* Firmware sets up some stuff, then `mret`s into the bootloader in
  S-mode.
* The "bootloader" (src/boot/riscv.s) parses the program headers in
  the kernel ELF file and sets up appropriate PTEs.
  * All of the virtual addresses must be in the upper 2 MiB. Easily
    achievable with a microkernel, maybe not with monolithic kernels but
    w/e.
  * The rest of the PTEs in PPN[2] are identity mappings as gigapages.
  * The bootloader sets up `satp` in Sv39 mode.
* The bootloader then jumps to the entry point of the kernel. Then the
  kernel can whatever else it needs to do without needing to deal with
  identity mapping itself without ELF information.

Because the kernel was originally designed with M-mode in mind, which
has VMA disabled by default even if `satp` is set, a lot of stuff broke.
Some hacks have been added to the kernel and the userspace libraries /
init to make it work again. It will be fixed properly in a later
commit, as some major reworks are needed.