Remove buddy allocator, replace with stack allocator
The buddy allocator is annoying to use with VMA due to the use of linked
lists. A stack structure doesn't use linked lists but is simply a big
array and thus doesn't require mapping every single page.

The stack allocator should have a tree structure as a backing store but
the design still needs work, hence only a stack is used at the moment.