Rework VMM to use hugepage windows for modifying trees.
I don't know what the canonical name for the technique is, so I dub it
"hugepage windows".

To update physical memory location pages they can be temporarily mapped
as hugepages. Two slots are dedicated to this in the root table
(HIGHMEM_A and B). This should be reasonably fast as the table is cached
and it only requires one load (instead of several for 4K pages).

Identity mappings are removed right after early boot.

The trap handler checks if sscratch is 0 now and jumps to an 'early
panic' handler which dumps scause/sepc/stval, which is a lot nicer than
wading through hundreds of thousands of QEMU logs.