Implement redesigned IPC, make libdux Rust only
As I forgot the _why_ for the new IPC during implementation I'd figure I
quickly explain it:

The old IPC could only be used chronologously in practice: you can't
skip or defer packets that are irrelevant in the current context.
There were a few tricks to make it possible, but they had a relatively
large cost as it involves either boxing or swapping entire packets.

To remedy this, the new IPC design features a "free slot stack" and two
ring buffers using slots. A single slot is merely 2 bytes large, so
swapping them is cheap. Incidentally, this approach also uses less
memory and should improve cache efficiency by reusing the same slots
frequently thanks to the stack.

While implementing the redesigned IPC in Rust it was clear that
duplicating the effort for other languages such as C would be
error-prone and wasteful. Instead, `dux_ar` has been added, which
produces an archive other languages can link statically against.

There's also a large number of bug fixes but I can't be arsed
documenting all of them. Just know that this entire OS is broken as
fuck.