Attempt to get external interrupts to work without panic.
There is a big problem: it seems we're unable to exit the handler when
an external interrupt has been triggered. There is probably a workaround
for this but there is also another issue:

The current implementation requires that one task receives & handles all
the interrupts. This means that either a) all the drivers are in that
task to improve performance or b) it forwards a message / notification
to other tasks, which will incur another context switch.

Moving forward, it is probably best to integrate the PLIC and any other
kinds of interrupt controllers in the kernel as it is a critical service
for pretty much all drivers. This makes the kernel les "µ"-y, but then
again, we didn't choose the design out of dogma.