Push notification arguments on stack, preserve registers.
By delegating the need to preserve registers to the userspace we can
save complexity in the kernel by needing another buffer to store
registers in.

This does require that we either use userspace CSRs or the stack to
store the arguments though. The stack is used for now, but this makes us
unable to handle stack overflows in userspace unless stacks can be
switched on the fly.

To avoid needing more branches, this stack could be dedicated to
handling exceptional conditions. That way notifications & interrupts
remain faster & more consistent.